refactor: Move UseLogin hook to hooks directory
